refactor!: [DX-1742] Rename NumberPicker to Stepper
