+++ optimus/lib/optimus.dart
@@ -60,6 +60,7 @@ export 'src/form/input_form_field.dart';
 export 'src/form/password_form_field.dart';
 export 'src/form/select_form_field.dart';
 export 'src/form/select_input_form_field.dart';
+export 'src/form/stepper_form_field.dart';
 export 'src/icon.dart';
 export 'src/icon_list.dart';
 export 'src/link/inline_link.dart';
@@ -70,7 +71,6 @@ export 'src/lists/list_tile.dart';
 export 'src/lists/nav_list_tile.dart';
 export 'src/loader/spinner.dart';
 export 'src/logo.dart';
-export 'src/number_picker/number_picker.dart';
 export 'src/progress_indicator/progress_indicator.dart';
 export 'src/progress_indicator/progress_indicator_compact.dart';
 export 'src/progress_indicator/progress_indicator_item.dart';
